Add new "drag and release" point selection
In addition to being able to press E to select your start and end points, you can now press and hold E to set an "anchor point" and then while continuing to hold E, look at another place and then release E and your start or end point will be placed in the middle of the "anchor point" and the place you released.

For example with a parking space, you press and hold E on one end of the edge, then release E on the other end, and you'll get a point right in the center.