Refactor formatting system to be generic and extensible
This creates an extensible formatting system that makes it easy to output coord-lines in whatever format you want, as long as you create a formatter for it. A formatter just requires 4 things: a name and three functions (onBeforeVectors, onVectors, onAfterVectors). Each of the functions should output the formatted string for that section.

Also this commit organizes the file structure a bit better.